requirements.txt
python-evtx
PyYAML
rich
